Add an "enabled" member to the MIB struct.  By default, all core-related MIBs are enabled by default.  The other MIBs, e.g. those for mod_tls, mod_sftp, and others, are disabled by default.
When mod_snmp initializes, it will check for those other modules.  If they
are loaded, then the MIBs related to those modules will be enabled.  This is
done so that e.g. snmpwalk will pull only those objects which are currently
enabled, rather than everything (and having blank/zero objects for modules
which are not loaded).  Without this, the SNMP manager cannot tell the
difference between "zero because module not loaded" and "zero because that
is reality".